refactor: move functionality into transcript_utilities.py
